[syft] address review changes
