+++ packages/syft/src/syft/service/worker/worker_image_service.py
@@ -1,12 +1,7 @@
 # stdlib
 import contextlib
-import json
-import os
-from typing import Iterable
-from typing import Iterator
 from typing import List
 from typing import Optional
-from typing import Tuple
 from typing import Union
 
 # third party
@@ -14,8 +9,6 @@ import docker
 import pydantic
 
 # relative
-from ...custom_worker.builder import CustomWorkerBuilder
-from ...custom_worker.builder import Image
 from ...custom_worker.config import DockerWorkerConfig
 from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
@@ -31,6 +24,8 @@ from ..service import service_method
 from ..user.user_roles import DATA_OWNER_ROLE_LEVEL
 from .image_registry import SyftImageRegistry
 from .image_registry_service import SyftImageRegistryService
+from .utils import docker_build
+from .utils import docker_push
 from .worker_image import SyftWorkerImage
 from .worker_image import SyftWorkerImageTag
 from .worker_image_stash import SyftWorkerImageStash
@@ -40,12 +35,10 @@ from .worker_image_stash import SyftWorkerImageStash
 class SyftWorkerImageService(AbstractService):
     store: DocumentStore
     stash: SyftWorkerImageStash
-    builder: CustomWorkerBuilder
 
     def __init__(self, store: DocumentStore) -> None:
         self.store = store
         self.stash = SyftWorkerImageStash(store=store)
-        self.builder = CustomWorkerBuilder()
 
     @service_method(
         path="worker_image.submit_dockerfile",
@@ -79,7 +72,6 @@ class SyftWorkerImageService(AbstractService):
         version: str = "latest",
         registry: Optional[UID] = None,
     ) -> Union[SyftSuccess, SyftError]:
-        worker_image: SyftWorkerImage = None
         registry_obj: Optional[SyftImageRegistry] = None
         image_registry_service: Optional[SyftImageRegistryService] = None
 
@@ -88,9 +80,9 @@ class SyftWorkerImageService(AbstractService):
             return SyftError(
                 message=f"Failed to get image for uid: {image}. Error: {result.err()}"
             )
-        worker_image = result.ok()
+        worker_image: SyftWorkerImage = result.ok()
 
-        if worker_image.built_on:
+        if worker_image.built_at:
             return SyftError(message=f"Image<{image}> is already built")
 
         if registry:
@@ -114,13 +106,13 @@ class SyftWorkerImageService(AbstractService):
 
         worker_image.image_tag = image_tag
 
-        result = self.docker_build(image=worker_image)
+        result = docker_build(image=worker_image)
 
         if isinstance(result, SyftError):
             return result
 
         (image, logs) = result
-        worker_image.built_on = DateTime.now()
+        worker_image.built_at = DateTime.now()
         worker_image.image_hash = image.id
 
         update_result = self.stash.update(context.credentials, obj=worker_image)
@@ -144,20 +136,19 @@ class SyftWorkerImageService(AbstractService):
         username: str = "",
         password: str = "",
     ) -> Union[SyftSuccess, SyftError]:
-        worker_image: SyftWorkerImage = None
         result = self.stash.get_by_uid(credentials=context.credentials, uid=image)
         if result.is_err():
             return SyftError(
                 message=f"Failed to get image for uid: {image}. Error: {result.err()}"
             )
-        worker_image = result.ok()
+        worker_image: SyftWorkerImage = result.ok()
 
-        if worker_image.built_on is None:
+        if worker_image.built_at is None:
             return SyftError(
                 message=f"Image {worker_image} is not built yet. Please build it first."
             )
 
-        result = self.docker_push(
+        result = docker_push(
             image=worker_image,
             username=username,
             password=password,
@@ -196,6 +187,7 @@ class SyftWorkerImageService(AbstractService):
         res = self.stash.get_by_uid(credentials=context.credentials, uid=uid)
         if res.is_err():
             return SyftError(message=f"{res.err()}")
+
         image: SyftWorkerImage = res.ok()
 
         if image and image.image_tag:
@@ -220,76 +212,6 @@ class SyftWorkerImageService(AbstractService):
         )
 
         return SyftSuccess(message=returned_message)
-
-    def docker_build(
-        self, image: SyftWorkerImage
-    ) -> Union[Tuple[Image, Iterable[str]], SyftError]:
-        try:
-            (image, logs) = self.builder.build_image(
-                config=image.config,
-                tag=image.image_tag.full_tag,
-            )
-            parsed_logs = self.parse_output(logs)
-            return (image, parsed_logs)
-        except docker.errors.APIError as e:
-            return SyftError(
-                message=f"Docker API error when building {image.image_tag}. Reason - {e}"
-            )
-        except docker.errors.DockerException as e:
-            return SyftError(
-                message=f"Docker exception when building {image.image_tag}. Reason - {e}"
-            )
-        except Exception as e:
-            return SyftError(
-                message=f"Unknown exception when building {image.image_tag}. Reason - {e}"
-            )
-
-    def docker_push(
-        self,
-        image: SyftWorkerImage,
-        username: str = "",
-        password: str = "",
-    ) -> List[str]:
-        try:
-            result = self.builder.push_image(
-                tag=image.image_tag.full_tag,
-                registry_url=image.image_tag.registry_host,
-                username=username,
-                password=password,
-            )
-
-            parsed_result = result.split(os.linesep)
-
-            if "error" in result:
-                result = SyftError(
-                    message=f"Failed to push {image.image_tag}. Logs - {parsed_result}"
-                )
-
-            return parsed_result
-        except docker.errors.APIError as e:
-            return SyftError(
-                message=f"Docker API error when pushing {image.image_tag}. {e}"
-            )
-        except docker.errors.DockerException as e:
-            return SyftError(
-                message=f"Docker exception when pushing {image.image_tag}. Reason - {e}"
-            )
-        except Exception as e:
-            return SyftError(
-                message=f"Unknown exception when pushing {image.image_tag}. Reason - {e}"
-            )
-
-    def parse_output(self, log_iterator: Iterator) -> str:
-        log = ""
-        for line in log_iterator:
-            for item in line.values():
-                if isinstance(item, str):
-                    log += item
-                elif isinstance(item, dict):
-                    log += json.dumps(item)
-                else:
-                    log += str(item)
-        return log
 
 
 TYPE_TO_SERVICE[SyftWorkerImage] = SyftWorkerImageService
